\name{ovarian}
\alias{ovarian}
\docType{data}
\title{ovarian
}
\description{
copy of ovarian from package 'survival'. Will disappear
}
\usage{data("ovarian")}
\format{
  A data frame with 26 observations on the following 6 variables.
  \describe{
    \item{\code{futime}}{a numeric vector}
    \item{\code{fustat}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{resid.ds}}{a numeric vector}
    \item{\code{rx}}{a numeric vector}
    \item{\code{ecog.ps}}{a numeric vector}
  }
}
\details{
  This copy is here since the package was rejected because the checking
  procedure did not find it in the package
}
\examples{
data(ovarian)
summary(ovarian)
}
\keyword{datasets}
