\name{relevance-internal}
%% \alias{i.def}
%% %% \alias{i.smooth}
%% %
%% \alias{is.formula}
%% \alias{\%nin\%}
%% %% \alias{print.mregr}
%% \alias{i.main}
%% \alias{u.true}
%% \alias{u.notfalse}
%% \alias{u.debug}
\alias{DB}
%% ---
\title{Internal objects of package relevance}
\description{
  Internal functions of package \pkg{relevance} that are exported, but
  should normally not be called by the user directly.
}
\usage{
DB(on=TRUE)
}
\arguments{
  \item{on}{
    0 or \code{FALSE} for switching debugging off,\cr
    1 or \code{TRUE} for switching debugging on,\cr
    2 for turning warnings into errors
  }
}
\author{\packageAuthor{relevance}}
\keyword{internal}
